Move TimelineEvent definition into timeline-types.ts.
