+++ src/plugins/timeline/timeline-types.ts
@@ -1,3 +1,12 @@
+import { DateTime } from "luxon";
+
 export const kTimelineTileType = "Timeline";
 
 export const kTimelineDefaultHeight = 320;
+
+export interface TimelineEvent {
+  index: number;
+  windowStart: DateTime;
+  windowEnd: DateTime;
+  eventType: string;
+}
